Add token tracking functions to user object
We tokens can be added and remove from a user object, they are
never serialuzed to json (the token itself that is) but only
represented as a ID to prevent them to be send outside of the
boundaries of the application presenting possible security
reists. after all tokens are plaintext passwords in the end!

On every load of an objects tokens are expiered this guarantees
that tokens are bounded even if many are added, unless those are
added with an infinity timeout - which would be stupid.